Move to backend-common
