+++ backend/native/backend-common/src/db/index.ts
@@ -2,4 +2,5 @@ export * from "./chats";
 export * from "./chatUpdateHistory";
 export * from "./enrichMessages";
 export * from "./friendship";
+export * from "./notifications";
 export * from "./user";
